+++ packages/suite/src/actions/wallet/constants/fiatRatesConstants.ts
@@ -2,7 +2,3 @@ export const RATE_UPDATE = '@rate/update';
 export const RATE_REMOVE = '@rate/remove';
 export const LAST_WEEK_RATES_UPDATE = '@rate/last-week-rate-update';
 export const TX_FIAT_RATE_UPDATE = '@rate/tx-fiat-rate-update';
-
-export const FETCH_COIN_LIST_START = '@rate/fetch-coin-list-start';
-export const FETCH_COIN_LIST_SUCCESS = '@rate/fetch-coin-list-success';
-export const FETCH_COIN_LIST_FAIL = '@rate/fetch-coin-list-fail';